requirements.txt
# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Configuration
pyyaml>=6.0

# Environment
python-dotenv>=1.0.0

# Async HTTP client (used by api.llm and api.translation)
httpx>=0.27.0

# Data model
pydantic>=2.0
